# lib/db/src/schema/dfm-findings.ts
/**
 * DFM Findings Table
 *
 * Stores Design for Manufacturability findings generated by the DFM
 * reviewer. Each finding belongs to an analysis session and includes
 * a category, severity, description, recommendation, and optional
 * references to related annotations.
 */
import { jsonb, pgTable, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

import { analysisSessions } from "./analysis-sessions";

export const dfmFindings = pgTable("dfm_findings", {
  id: text("id").primaryKey(),
  sessionId: text("session_id")
    .notNull()
    .references(() => analysisSessions.id, { onDelete: "cascade" }),
  category: text("category", {
    enum: [
      "over_tolerancing",
      "missing_tolerance",
      "datum_scheme_completeness",
      "surface_finish_consistency",
      "general",
    ],
  }).notNull(),
  severity: text("severity", {
    enum: ["error", "warning", "info"],
  }).notNull(),
  description: text("description").notNull(),
  recommendation: text("recommendation").notNull(),
  relatedAnnotationIds: jsonb("related_annotation_ids")
    .$type<string[]>()
    .default([]),
  createdAt: timestamp("created_at", { withTimezone: true })
    .defaultNow()
    .notNull(),
});

/** Zod schema for inserting a new DFM finding (auto-generated fields omitted). */
export const insertDfmFindingSchema = createInsertSchema(dfmFindings).omit({
  createdAt: true,
});

export type DfmFinding = typeof dfmFindings.$inferSelect;
export type InsertDfmFinding = z.infer<typeof insertDfmFindingSchema>;